+++ docs/source/conf.py
@@ -16,7 +16,8 @@ import os
 import sys
 # sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('../../'))
-sys.path.insert(0, os.path.abspath('../../libs'))
+# sys.path.insert(0, os.path.abspath('../../libs'))
+from kwarwp import __version__
 
 
 # -- Project information -----------------------------------------------------
@@ -26,9 +27,9 @@ copyright = '2020, Carlo E. T. Oliveira'
 author = 'Carlo E. T. Oliveira'
 
 # The short X.Y version
-version = '20.07.1'
+version = __version__
 # The full version, including alpha/beta/rc tags
-release = '20.07.1'
+release = __version__
 
 
 # -- General configuration ---------------------------------------------------